+++ src/main/java/tkcy/simpleaddon/api/items/toolitem/TKCYSAToolClasses.java
@@ -1,0 +1,6 @@
+package tkcy.simpleaddon.api.items.toolitem;
+
+public class TKCYSAToolClasses {
+
+    public static final String SOLDERING_IRON = "soldering_iron";
+}