+++ src/main/java/tkcy/simpleaddon/api/unification/flags/FlagsAddition.java
@@ -27,6 +27,10 @@ public class FlagsAddition {
 
         GENERATE_ALL_NO_UNIF.forEach(flag -> Carbon.addFlags(flag));
 
+        Steel.addFlags(GENERATE_SOLDERING_IRON_TIP);
+
+        addAllAlloyFlags(TinAlloy);
+
         electrodeMaterials.forEach(FlagsAddition::addElectrodeFlag);
         alloysMaterials.forEach(FlagsAddition::addAlloyFlag);
         getLvAcceptedRubberMaterials().forEach(FlagsAddition::addRingAndPlateFlags);
@@ -42,5 +46,9 @@ public class FlagsAddition {
 
     public static void addRingAndPlateFlags(Material material) {
         material.addFlags(GENERATE_RING, GENERATE_PLATE);
+    }
+
+    public static void addAllAlloyFlags(Material material) {
+        GENERATE_ALL.forEach(material::addFlags);
     }
 }