+++ src/main/java/tkcy/simpleaddon/api/unification/ore/OrePrefixAdditions.java
@@ -1,0 +1,10 @@
+package tkcy.simpleaddon.api.unification.ore;
+
+import gregtech.api.unification.material.Materials;
+
+public class OrePrefixAdditions {
+
+    public static void init() {
+        TKCYSAOrePrefix.toolTipSolderingIron.processOreRegistration(Materials.Steel);
+    }
+}