+++ src/main/java/tkcy/simpleaddon/api/unification/ore/OrePrefixRegistry.java
@@ -9,8 +9,10 @@ public class OrePrefixRegistry {
         addOrePrefix(TKCYSAOrePrefix.electrode);
         addOrePrefix(TKCYSAOrePrefix.anode);
         addOrePrefix(TKCYSAOrePrefix.cathode);
+        addOrePrefix(curvedPlate);
 
         registerComponentsOres();
+        registerToolOres();
     }
 
     private static void registerComponentsOres() {
@@ -22,5 +24,9 @@ public class OrePrefixRegistry {
         addOrePrefix(luvComponents);
         addOrePrefix(zpmComponents);
         addOrePrefix(uvComponents);
+    }
+
+    private static void registerToolOres() {
+        addOrePrefix(toolTipSolderingIron);
     }
 }