+++ src/main/java/tkcy/simpleaddon/api/unification/ore/TKCYSAOrePrefix.java
@@ -5,6 +5,7 @@ import static gregtech.api.unification.ore.OrePrefix.Flags.ENABLE_UNIFICATION;
 import static gregtech.api.unification.ore.OrePrefix.Flags.SELF_REFERENCING;
 
 import gregtech.api.unification.material.MarkerMaterials;
+import gregtech.api.unification.material.info.MaterialFlags;
 import gregtech.api.unification.ore.OrePrefix;
 
 import tkcy.simpleaddon.api.unification.flags.TKCYSAMaterialFlags;
@@ -23,6 +24,11 @@ public class TKCYSAOrePrefix {
             ENABLE_UNIFICATION,
             mat -> mat.hasFlag(TKCYSAMaterialFlags.GENERATE_ELECTRODES));
 
+    public static final OrePrefix curvedPlate = new OrePrefix("curved_plate", M, null,
+            TKCYSAMaterialIconType.curvedPlate,
+            ENABLE_UNIFICATION,
+            mat -> mat.hasFlag(MaterialFlags.GENERATE_ROTOR));
+
     // Components
     public static final OrePrefix lvComponents = new OrePrefix("lvComponents", -1, MarkerMaterials.Empty,
             TKCYSAMaterialIconType.lvComponents, SELF_REFERENCING, null);
@@ -40,4 +46,9 @@ public class TKCYSAOrePrefix {
             TKCYSAMaterialIconType.zpmComponents, SELF_REFERENCING, null);
     public static final OrePrefix uvComponents = new OrePrefix("uvComponents", -1, MarkerMaterials.Empty,
             TKCYSAMaterialIconType.uvComponents, SELF_REFERENCING, null);
+
+    // Tools
+    public static final OrePrefix toolTipSolderingIron = new OrePrefix("toolTipSolderingIron", 1, null,
+            TKCYSAMaterialIconType.toolTipSolderingIron, ENABLE_UNIFICATION,
+            mat -> mat.hasFlag(TKCYSAMaterialFlags.GENERATE_SOLDERING_IRON_TIP));
 }