+++ src/main/java/tkcy/simpleaddon/common/ClientProxy.java
@@ -5,7 +5,9 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.relauncher.Side;
 
+import tkcy.simpleaddon.api.render.TKCYSATextures;
 import tkcy.simpleaddon.common.block.TKCYSAMetaBlocks;
+import tkcy.simpleaddon.common.item.TKCYSAToolItems;
 
 @Mod.EventBusSubscriber(Side.CLIENT)
 public class ClientProxy extends CommonProxy {
@@ -13,10 +15,21 @@ public class ClientProxy extends CommonProxy {
     @Override
     public void preLoad() {
         super.preLoad();
+        TKCYSATextures.preInit();
     }
 
     @SubscribeEvent
     public static void registerModels(ModelRegistryEvent event) {
         TKCYSAMetaBlocks.registerItemModels();
+        TKCYSAToolItems.registerModels();
+    }
+
+    @Override
+    public void onLoad() {
+        registerColors();
+    }
+
+    public static void registerColors() {
+        TKCYSAToolItems.registerColors();
     }
 }